modularized the lab into components + renamed Footer to footer
